requirements.txt
django==4.2
environs==9.5.0
Pillow==9.5.0
django-rosetta==0.9.8
django_parler==2.3
